Convert all ompi_bitmap_t and ompi_pointer_array_t indexing from size_t to int.  This is in recognition of the fact that these two classes are primarily used for fortran<-->c convertsion of various entities (attributes, MPI objects), and MPI defines that an int must be used to hold MPI fortran handles.  Hence, why use size_t internally?   See the comment in src/class/ompi_pointer_array.h for a better description.
Mixed in a few attribute fixes in this commit as well.

This commit, therefore:
- converts the indexing from size_t to int
- changes all locations in the code that I could find that used size_t
  as the interface to ompi_bitmap_t or ompi_pointer_array_t
- convert to use OMPI_FINT_2_INT / OMPI_INT_2_FINT macros in various
  src/mpi/c/*f2c.c and src/mpi/c/*c2f.c files
- check the return code of ompi_comm_free() to ensure that everything
  worked properly before returning MPI_SUCCESS
- unified all src/mpi/c/*f2c.c functions (i.e., they all do the same
  thing)
- tie up some loose ends w.r.t. MPI_Request handling in fortran; set
  the req_f_to_c_index to MPI_UNDEFINED when it does not have a fortran
  index
- still need to add a configure test to find OMPI_FORTRAN_HANDLE_MAX
  for ompi_bitmap.c and ompi_pointer_array.c (hard-wired to INT_MAX
  right now)
- re-organized, consolidated, and unified some ompi_pointer_array.c
  code -- fixed a few minor bugs w.r.t. lowest_free (could have left
  some unintentional holes in the array)

This commit was SVN r3302.

+++ src/class/ompi_bitmap.h
@@ -13,14 +13,20 @@
  *  bit is set -- so the valid functions are set_bit and
  *  find_and_set_bit. Other functions like clear, if passed a bit
  *  outside the initialized range will result in an error.
+ *
+ * Since these bitmaps are only used to track fortran handles (which
+ * MPI defines to be int's), it is assumed that we can never have more
+ * than OMPI_FORTRAN_HANDLE_MAX (which is min(INT_MAX, fortran
+ * INTEGER max)).
  */
 
 #ifndef OMPI_BITMAP_H
 #define OMPI_BITMAP_H
 
+#include "ompi_config.h"
+
 #include <string.h>
 
-#include "ompi_config.h"
 #include "include/types.h"
 #include "class/ompi_object.h"
 
@@ -30,8 +36,8 @@ extern "C" {
 struct ompi_bitmap_t {
     ompi_object_t super; /**< Subclass of ompi_object_t */
     unsigned char *bitmap; /**< The actual bitmap array of characters */
-    size_t array_size;  /**< The actual array size that maintains the bitmap */
-    size_t legal_numbits; /**< The number of bits which are legal (the
+    int array_size;  /**< The actual array size that maintains the bitmap */
+    int legal_numbits; /**< The number of bits which are legal (the
 			    actual bitmap may contain more bits, since
 			    it needs to be rounded to the nearest
 			    char  */
@@ -50,7 +56,7 @@ OMPI_DECLSPEC OBJ_CLASS_DECLARATION(ompi_bitmap_t);
  * @return OMPI error code or success
  *
  */
-OMPI_DECLSPEC int ompi_bitmap_init (ompi_bitmap_t *bm, size_t size);
+OMPI_DECLSPEC int ompi_bitmap_init (ompi_bitmap_t *bm, int size);
 
 
 /**
@@ -63,7 +69,7 @@ OMPI_DECLSPEC int ompi_bitmap_init (ompi_bitmap_t *bm, size_t size);
  * @return OMPI error code or success
  *
  */
-OMPI_DECLSPEC int ompi_bitmap_set_bit(ompi_bitmap_t *bm, size_t bit); 
+OMPI_DECLSPEC int ompi_bitmap_set_bit(ompi_bitmap_t *bm, int bit); 
 
 
 /**
@@ -75,7 +81,7 @@ OMPI_DECLSPEC int ompi_bitmap_set_bit(ompi_bitmap_t *bm, size_t bit);
  * @return OMPI error code if the bit is out of range, else success
  *
  */
-OMPI_DECLSPEC int ompi_bitmap_clear_bit(ompi_bitmap_t *bm, size_t bit);
+OMPI_DECLSPEC int ompi_bitmap_clear_bit(ompi_bitmap_t *bm, int bit);
 
 
 /**
@@ -88,7 +94,7 @@ OMPI_DECLSPEC int ompi_bitmap_clear_bit(ompi_bitmap_t *bm, size_t bit);
   *         0 if the bit is not set
   *
   */
-OMPI_DECLSPEC int ompi_bitmap_is_set_bit(ompi_bitmap_t *bm, size_t bit);
+OMPI_DECLSPEC int ompi_bitmap_is_set_bit(ompi_bitmap_t *bm, int bit);
 
 
 /**
@@ -100,7 +106,7 @@ OMPI_DECLSPEC int ompi_bitmap_is_set_bit(ompi_bitmap_t *bm, size_t bit);
  * @return err        OMPI_SUCCESS on success
  */
 OMPI_DECLSPEC int ompi_bitmap_find_and_set_first_unset_bit(ompi_bitmap_t *bm, 
-                                             size_t *position); 
+                                                           int *position); 
 
 
 /**
@@ -130,7 +136,7 @@ OMPI_DECLSPEC int ompi_bitmap_set_all_bits(ompi_bitmap_t *bm);
  * @return OMPI error code if bm is NULL
  *
  */
-static inline size_t ompi_bitmap_size(ompi_bitmap_t *bm)
+static inline int ompi_bitmap_size(ompi_bitmap_t *bm)
 {
     return (NULL == bm) ? 0 : bm->legal_numbits;
 }
